requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
asgiref==3.2.10
async-exit-stack==1.0.1
async-generator==1.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
dnspython==2.0.0
email-validator==1.1.1
fastapi==0.61.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
orjson==3.3.1
promise==2.3
psycopg2-binary==2.8.5
pydantic==1.6.1
PyJWT==1.7.1
python-dotenv==0.14.0
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
starlette==0.13.6
ujson==3.1.0
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
